cloud-tasks/cloud-function-to-composer/requirements.txt
#airflow
apache-airflow[gcp]==1.10.12

#cloud-function
google-cloud-tasks